refactor: hook 분리
[#22]
